Add files via upload
The first file is a simple employee table with the employee ID, name, department, position, salary and the output. 
The second file I added the hire_date but came up with an error. I realized that I didn't up the single marks around the email address. 
The third file the output field resulted in an error. After double checking again, I didn't have a value for employee_id.
The last file is a successful output.